LPD-89430 move to subscriptionentry service
